+++ src/app/community/packages/[ecosystem]/[...nameAndVersion]/_components/stats-cards/license-stats-card.tsx
