+++ src/app/community/packages/[ecosystem]/[...nameAndVersion]/_components/stats-cards/openssf-score-card.tsx
