+++ src/app/community/packages/[ecosystem]/[...nameAndVersion]/_components/stats-cards/stats-cards.tsx
