+++ src/app/community/packages/[ecosystem]/[...nameAndVersion]/_components/stats-cards/vulnerability-stats-card.tsx
